equivalent tiles: refactored tile tag script
